Moves BaseDefaultValidator to base
It was causing circular imports otherwise.
